# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Voice Audit API",
    "description": "\nAI Voice Audit API provides a robust set of endpoints for automating the quality assurance process in call centers using advanced AI models.\n\n## Key Features\n- **Asynchronous Audio Processing**: Upload call recordings for transcription and AI-powered evaluation.\n- **Dynamic LLM Selection**: Configure specific AI models per Line of Business (LOB).\n- **Multi-Tenancy**: Secure data isolation and tenant-specific configurations.\n- **Advanced Analytics**: Real-time dashboards and exportable reports.\n\n## Authentication\nThis API uses **JWT tokens** for secure authentication. \n1. Obtain tokens via the `/api/auth/login` endpoint.\n2. Include the access token in the `Authorization` header for all protected endpoints:\n   `Authorization: Bearer <your_access_token>`\n",
    "version": "1.0.0"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Authenticate User",
        "description": "Authenticate a user with email and password to obtain JWT tokens.\n\n## Example curl:\n```bash\ncurl -X POST \"http://localhost:5000/api/auth/login\"          -H \"Content-Type: application/json\"          -d '{\"email\": \"admin@acme.com\", \"password\": \"SecurePass456!\"}'\n```",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required or invalid token",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Could not validate credentials"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh Access Token",
        "description": "Obtain a new access token using a valid refresh token.\n\n## Example curl:\n```bash\ncurl -X POST \"http://localhost:5000/api/auth/refresh\"          -H \"Content-Type: application/json\"          -d '{\"refresh_token\": \"eyJhbGci...\"}'\n```",
        "operationId": "refresh_api_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required or invalid token",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Could not validate credentials"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout User",
        "description": "Invalidate the current access token by adding it to the blacklist.\n\n## Example curl:\n```bash\ncurl -X POST \"http://localhost:5000/api/auth/logout\"          -H \"Authorization: Bearer <access_token>\"\n```",
        "operationId": "logout_api_auth_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required or invalid token",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Could not validate credentials"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/dashboard/trends": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Score Trends",
        "description": "Get average score trends over time, grouped by the specified period.",
        "operationId": "get_trends_api_dashboard_trends_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Aggregation period: week or month",
              "default": "week",
              "title": "Period"
            },
            "description": "Aggregation period: week or month"
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "lob_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lob Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required or invalid token",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Could not validate credentials"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/ctq-distribution": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get CTQ Distribution",
        "description": "Get the pass/fail distribution for Critical to Quality (CTQ) criteria.",
        "operationId": "get_ctq_distribution_api_dashboard_ctq_distribution_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "lob_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lob Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required or invalid token",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Could not validate credentials"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/topic-trends": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Topic Frequency Trends",
        "description": "Get topic frequency trends over time, allowing identification of common call drivers.",
        "operationId": "get_topic_trends_api_dashboard_topic_trends_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Aggregation period: week or month",
              "default": "week",
              "title": "Period"
            },
            "description": "Aggregation period: week or month"
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "lob_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lob Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required or invalid token",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Could not validate credentials"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RefreshTokenRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token",
            "description": "The refresh token obtained during login",
            "examples": [
              "eyJhbGci..."
            ]
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshTokenRequest",
        "example": {
          "refresh_token": "eyJhbGci..."
        }
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token",
            "description": "JWT access token",
            "examples": [
              "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            ]
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token",
            "description": "JWT refresh token used to obtain new access tokens",
            "examples": [
              "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            ]
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "description": "The type of token returned",
            "default": "bearer",
            "examples": [
              "bearer"
            ]
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "user"
        ],
        "title": "TokenResponse",
        "example": {
          "access_token": "eyJhbGci...",
          "refresh_token": "eyJhbGci...",
          "token_type": "bearer",
          "user": {
            "created_at": "2024-01-01T12:00:00",
            "email": "admin@acme.com",
            "id": 1,
            "role": "admin"
          }
        }
      },
      "UserLogin": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "User's registered email",
            "examples": [
              "admin@acme.com"
            ]
          },
          "password": {
            "type": "string",
            "title": "Password",
            "description": "User's secret password",
            "examples": [
              "SecurePass456!"
            ]
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "UserLogin"
      },
      "UserResponse": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "The user's email address",
            "examples": [
              "user@company.com"
            ]
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "The role assigned to the user",
            "default": "admin",
            "examples": [
              "admin",
              "qa_manager",
              "analyst"
            ]
          },
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "The unique identifier for the user",
            "examples": [
              1
            ]
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At",
            "description": "The timestamp when the user was created"
          }
        },
        "type": "object",
        "required": [
          "email",
          "id"
        ],
        "title": "UserResponse",
        "example": {
          "created_at": "2024-01-01T12:00:00",
          "email": "admin@acme.com",
          "id": 1,
          "role": "admin"
        }
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/auth/login"
          }
        }
      }
    }
  },
...
}